+++ tests/gentype_tests/typescript-react-example/src/Core.gen.tsx
@@ -68,3 +68,7 @@ export const weakset1: (x:WeakSet<number[]>) => WeakSet<number[]> = CoreJS.weaks
 export const option0: (x:(undefined | string)) => (undefined | string) = CoreJS.option0 as any;
 
 export const option1: (x:(undefined | variant)) => (undefined | variant) = CoreJS.option1 as any;
+
+export const jsonEncodeString1: unknown = CoreJS.jsonEncodeString1 as any;
+
+export const jsonEncodeString2: unknown = CoreJS.jsonEncodeString2 as any;